requirements.txt
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.9
pyOpenSSL==20.0.1
pyscard==2.0.0
PySimpleGUI==4.33.0
six==1.15.0
